deleted useless files, refactored Controllers to be in the correct packaging according to Clean Architecture rules.
