+++ csc207/src/main/java/interface_adapter/HomeController.java
@@ -1,0 +1,14 @@
+package interface_adapter;
+
+import data_access.UserData;
+
+public class HomeController {
+    private UserData userData;
+    public HomeController(UserData userData) {
+        this.userData = userData;
+    }
+
+    public void execute() {
+
+    }
+}
