--- csc207/src/main/java/interface_adapter/home/HomeState.java
@@ -1,19 +1,0 @@
-package interface_adapter.home;
-
-/**
- * The state for the Home View Model.
- */
-public class HomeState {
-    private float balance = 0;
-    private float income = 0;
-    private float expenses = 0;
-
-    @Override
-    public String toString() {
-        return "HomeState{"
-                + "NetBalance'" + balance + '\''
-                + ", Income'" + income + '\''
-                + ", Expenses'" + expenses + '\''
-                + '}';
-    }
-}
